+++ scrollpointclick.AI/src/support/helpers.js
@@ -90,7 +90,7 @@ export async function adjustPreferences() {
 }
 
 export function removeEntry(heading: string) {
-  logError(pluginJson, `\n\n----- ----- -----\n${heading}\n\n---- ----- ---- \n\n`)
+  logDebug(pluginJson, `\n\n----- ----- -----\n${heading}\n\n---- ----- ---- \n\n`)
   const paraBeforeDelete = Editor.paragraphs.find((p) => p.content === heading)
   if (paraBeforeDelete) {
     const contentRange = paraBeforeDelete.contentRange
@@ -101,7 +101,7 @@ export function removeEntry(heading: string) {
 }
 
 export function scrollToEntry(heading: string, deleteItem?: bool = false, foldHeading?: bool = false) {
-  logError(pluginJson, `\n\n----- ----- -----\n${heading}\n\n${deleteItem}\n\n---- ----- ---- \n\n`)
+  logDebug(pluginJson, `\n\n----- ----- -----\n${heading}\n\n${deleteItem}\n\n---- ----- ---- \n\n`)
   const selectedHeading = Editor.paragraphs.find((p) => p.content === heading)
   if (selectedHeading) {
     const contentRange = selectedHeading.contentRange
@@ -109,7 +109,7 @@ export function scrollToEntry(heading: string, deleteItem?: bool = false, foldHe
     if (deleteItem) {
       
       firstCharacter = contentRange.start - 1 // back up one character
-      logError(pluginJson, `\n\n----- ----- -----\n${firstCharacter}\n\n---- ----- ---- \n\n`)
+      logDebug(pluginJson, `\n\n----- ----- -----\n${firstCharacter}\n\n---- ----- ---- \n\n`)
       // removeContentUnderHeading(Editor, heading, true, false)
       removeEntry(heading)
     } else {